split fuzzer logic out from app
